Add new entries on property object
Introduce two new entries on property object:

* `filePath`: the file path where the property comes from.
* `isSource`: describe if the property is a source or an include.

To set these two properties, the method `combineJSON` has now one
more parameter `source`. This is a boolean which
describes if json files are "source" or "include".

By the way, `filePath` and `isSource` properties are removed
from the original object properties (like it was in file).