fix(export platform): fixing infinite loop when there are reference errors (#531)
When there were property reference errors (could not find reference or circular reference) Style Dictionary would get caught in an infinite loop. To fix this, when Style Dictionary tries to resolve an erroneous reference (circular or not found), it replaces the reference with `undefined`. This is different than previous behavior where Style Dictionary would leave the erroneous reference. In both cases, Style Dictionary throws an error and fails, so this should not have a discernible effect. We also feel replacing the reference with `undefined` makes logical sense as that reference is undefined because it cannot be resolved.